moved entities from 'model' to 'entities'
