+++ src/common/locked-asset/entities/locked.tokens.interface.ts
