+++ src/common/locked-asset/entities/unlock.fields.interface.ts
